+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## master
+* Remove `bindOnce`, `renderAsync` & `renderAsyncWrapper`;
 * [#18](https://github.com/sham-ui/sham-ui/issues/18) remove `@handler`
 * [#20](https://github.com/sham-ui/sham-ui/issues/20) inject as getter
 * Remove options conflict resolver